+++ R/Connection.R
@@ -123,10 +123,10 @@ setMethod(
   c("OdbcConnection", "Id"),
   function(conn, name, column_name = NULL) {
 
-    odbcConnectionColumns(conn,
-      name = name@name[["table"]],
-      catalog_name = name@name[["catalog"]],
-      schema_name = name@name[["schema"]],
+    connection_sql_columns(conn@ptr,
+      table_name = id_field(name, "table"),
+      catalog_name = id_field(name, "catalog"),
+      schema_name = id_field(name, "schema"),
       column_name = column_name)
   }
 )