+++ NEWS.md
@@ -1,5 +1,17 @@
 # odbc (development version)
 
+* odbc can now be compiled again with Rtools35 (gcc 4.9.3) (#383)
+* `invalid descriptor` issues due to out of order retrieval of long columns are now avoided by unbinding any nanodbc buffer past the long column. Performance for the unbound columns in these cases will be reduced, but the retrieval will work (@detule, #381)
+* `dbWriteTable()` now handles `data.table::IDate()` objects (#388)
+* `odbcConnectionColumns()` now works as intended with `DBI::Id()` objects (#389)
+* Oracle connections now support `Date` and `POSIXct` types via `DATE` and `TIMESTAMP` data types (#324, #349, #350)
+* Oracle connections now use VARCHAR2 rather than VARCHAR, as recommended by Oracle's documentation (#189)
+* `dbWriteTable(field.types=)` now issues a warning rather than an error for missing columns (#342)
+* New `dbAppendTable()` method for OdbcConnection objects (#335)
+* `dbWriteTable()` and `dbBind()` now default to a `batch_rows` of `NA`, which sets the batch size to be the length of the input.
+  This avoids problems with drivers that don't support batch sizes larger than the input size.
+  To restore the behavior prior to this release pass `batch_rows = 1024` or set `options(odbc.batch_rows = 1024)` (#391).
+
 # odbc 1.2.3
 
 * `dbWriteTable()` now executes immediately, which fixes issues with temporary tables and the FreeTDS SQL Server driver (@krlmlr).